Create Doing HTTP requests
This is a short introduction to requests, a Python module that make grabbing content from the web quite easy.

Requests is NOT in the python distribution, but is installed along Anaconda.

If it's not installed, try: python3 -m pip install --user requests.

Now, your exercice will just have to GET the content of the page https://api.github.com/users/python, and print it.

In case your computer is not connected to the internet, your program should simply print No internet connectivity. on the standard output.

References
Requests: https://requests.readthedocs.io
Example
$ python solution.py
{
  "login": "python",
  "id": 1525981,
  ...
Oh Damn ! The wifi is down !

$ python solution.py
No internet connectivity.