Create Break a Safe
You found a safe. There's a numeric keypad on it...

The keypad is dirty, thanks to this you can easily spot that numbers 1, 5, and 8 are frequently used. It's so clear that you guess no other digit is in use in this code.

Write a script printing all combinations of 4 digits that use all of those used digits.

Output one combination per line, separate digits using spaces, so your output should look like:

1 1 5 8
1 5 5 8
1 5 8 8
...